requirements.txt
flask
numpy
pandas
matplotlib
tensorflow
seaborn
sklearnclear